+++ src/raft/append_entries.h
@@ -179,7 +179,7 @@ inline nlohmann::json AppendEntries::ToJson() {
   j["pre_log_term"] = pre_log_term;
   j["commit_index"] = commit_index;
   j["entry_count"] = entries.size();
-  for (int32_t i = 0; i < entries.size(); ++i) {
+  for (size_t i = 0; i < entries.size(); ++i) {
     j["entries"][i] = entries[i].ToJson();
   }
   j["this"] = PointerToHexStr(this);
@@ -195,7 +195,7 @@ inline nlohmann::json AppendEntries::ToJsonTiny() {
   j["ptm"] = pre_log_term;
   j["cmt"] = commit_index;
   j["cnt"] = entries.size();
-  for (int32_t i = 0; i < entries.size(); ++i) {
+  for (size_t i = 0; i < entries.size(); ++i) {
     j["etr"][i] = entries[i].ToJsonTiny();
   }
   j["ts"] = PointerToHexStr(this);
